// src/commands-slash/Ptu/components/fakemon/actionRowBuilders/moves/types.ts
export enum FakemonMovesButtonCustomIds
{
    AddLevelUpMoves = 'Add Level Up Moves',
    AddEggMoves = 'Add Egg Moves',
    AddTmHmMoves = 'Add TM/HM Moves',
    AddTutorMoves = 'Add Tutor Moves',
}

export enum FakemonMovesStringSelectCustomIds
{
    EditLevelUpMoves = '✏️ Edit Level Up Moves',
    RemoveLevelUpMoves = '🗑️ Remove Level Up Moves',
    RemoveEggMoves = '🗑️ Remove Egg Moves',
    RemoveTmHmMoves = '🗑️ Remove TM/HM Moves',
    RemoveTutorMoves = '🗑️ Remove Tutor Moves',
}